backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
google-genai==0.3.0
python-dotenv==1.0.0
PyPDF2==3.0.1
numpy==1.24.3
pandas==2.1.4
